+++ packages/rune-components/src/menu/index.ts
@@ -1,4 +1,36 @@
 export { default as Item } from './item/Item.svelte';
 export { default as Popup } from './popup/Popup.svelte';
-export { default as Trigger } from './trigger/Trigger.svelte';
 export { default as Root } from './root/Root.svelte';
+export { default as Separator } from './separator/Separator.svelte';
+export { default as Trigger } from './trigger/Trigger.svelte';
+
+export { getRootContext, setRootContext } from './context.js';
+
+export {
+	createItem,
+	type TItem,
+	type CreateItemConfig,
+} from './item/createItem.svelte.js';
+
+export {
+	createPopup,
+	type TPopup,
+	type CreatePopupConfig,
+} from './popup/createPopup.svelte.js';
+
+export {
+	createRoot,
+	type TRoot,
+	type CreateRootConfig,
+} from './root/createRoot.svelte.js';
+
+export {
+	createSeparator,
+	type TSeparator,
+} from './separator/createSeparator.svelte.js';
+
+export {
+	createTrigger,
+	type TTrigger,
+	type CreateTriggerConfig,
+} from './trigger/createTrigger.svelte.js';