+++ packages/rune-components/src/menu/separator/createSeparator.svelte.ts
@@ -1,3 +1,5 @@
+export type TSeparator = ReturnType<typeof createSeparator>;
+
 export function createSeparator() {
 	return {
 		props: {